[Bronze V] Title: 이상한 기호, Time: 176 ms, Memory: 17740 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -1,14 +1,14 @@
 import java.util.Scanner;
 
 class Main {
-    private static long calculate(int a, int b) {
+    private static long calculate(long a, long b) {
         return (a + b) * (a - b);
     }
 
     public static void main(String[] args) {
         final Scanner sc = new Scanner(System.in);
-        final int a = sc.nextInt();
-        final int b = sc.nextInt();
+        final long a = sc.nextLong();
+        final long b = sc.nextLong();
 
         final long result = calculate(a, b);
         System.out.println(result);
